Выполнена лабораторная 1
