+++ homework01/rsa.py
@@ -1,86 +1,28 @@
-import random
-from typing import Tuple
-
-
-def is_prime(n: int) -> bool:
+def encrypt_vigenere(plaintext: str, keyword: str) -> str:
     """
-    >>> is_prime(2)
-    True
-    >>> is_prime(11)
-    True
-    >>> is_prime(8)
-    False
+    Encrypts plaintext using a Vigenere cipher.
+    >>> encrypt_vigenere("PYTHON", "A")
+    'PYTHON'
+    >>> encrypt_vigenere("python", "a")
+    'python'
+    >>> encrypt_vigenere("ATTACKATDAWN", "LEMON")
+    'LXFOPVEFRNHR'
     """
-    if n < 2:
-        return False
-    if n == 2:
-        return True
-    if n % 2 == 0:
-        return False
-
-    # Проверяем делители от 3 до sqrt(n)
-    i = 3
-    while i * i <= n:
-        if n % i == 0:
-            return False
-        i += 2
-    return True
+    ciphertext = ""
+    # PUT YOUR CODE HERE
+    return ciphertext
 
 
-def gcd(a: int, b: int) -> int:
+def decrypt_vigenere(ciphertext: str, keyword: str) -> str:
     """
-    >>> gcd(12, 15)
-    3
-    >>> gcd(3, 7)
-    1
+    Decrypts a ciphertext using a Vigenere cipher.
+    >>> decrypt_vigenere("PYTHON", "A")
+    'PYTHON'
+    >>> decrypt_vigenere("python", "a")
+    'python'
+    >>> decrypt_vigenere("LXFOPVEFRNHR", "LEMON")
+    'ATTACKATDAWN'
     """
-    while b != 0:
-        a, b = b, a % b
-    return a
-
-
-def multiplicative_inverse(e: int, phi: int) -> int:
-    """
-    >>> multiplicative_inverse(7, 40)
-    23
-    """
-    # Используем расширенный алгоритм Евклида
-    # Находим x такое, что (e * x) % phi = 1
-    a, b, u = 0, phi, 1
-    while e > 0:
-        q = b // e
-        r = b % e
-        m = a - u * q
-        b, e, a, u = e, r, u, m
-    if b == 1:
-        return a % phi
-    else:
-        raise ValueError("Multiplicative inverse does not exist")
-
-
-def generate_keypair(p: int, q: int) -> Tuple[Tuple[int, int], Tuple[int, int]]:
-    if not (is_prime(p) and is_prime(q)):
-        raise ValueError("Both numbers must be prime.")
-    elif p == q:
-        raise ValueError("p and q cannot be equal")
-
-    # n = pq
-    n = p * q
-
-    # phi = (p-1)(q-1)
-    phi = (p - 1) * (q - 1)
-
-    # Choose an integer e such that e and phi(n) are coprime
-    e = random.randrange(1, phi)
-
-    # Use Euclid's Algorithm to verify that e and phi(n) are comprime
-    g = gcd(e, phi)
-    while g != 1:
-        e = random.randrange(1, phi)
-        g = gcd(e, phi)
-
-    # Use Extended Euclid's Algorithm to generate the private key
-    d = multiplicative_inverse(e, phi)
-    # Return public and private keypair
-    # Public key is (e, n) and private key is (d, n)
-    return ((e, n), (d, n))
+    plaintext = ""
+    # PUT YOUR CODE HERE
+    return plaintext
